+++ packages/cluster/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/cluster
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.3.0
+
 ## 0.5.0
 
 ### Minor Changes