+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/logger
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.3.0
+
 ## 0.1.1
 
 ### Patch Changes