+++ packages/provider-health/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/provider-health
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.3.0
+
 ## 0.3.1
 
 ### Patch Changes