+++ packages/provider-observability/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/provider-observability
 
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.3.0
+
 ## 0.4.1
 
 ### Patch Changes