+++ packages/runtime-distributed/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @shardix/runtime-distributed
 
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.3.0
+  - @shardix/core@0.3.0
+  - @shardix/cluster@0.5.1
+  - @shardix/ipc@0.4.2
+
 ## 0.4.1
 
 ### Patch Changes